pyproject.toml
[project]
name = "formula-one"
version = "0.1.0"
description = "A python module aiming to analyse and present statistics on Formula 1 teams and drivers performance."
readme = "README.md"
authors = [{ name = "Edwin Gelebart", email = "edwin.gelebart@gmail.com" }]

requires-python = ">=3.12"

dependencies = [
    "fastf1>=3.6.0",
    "gspread>=6.2.1",
    "pydantic>=2.11.7",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "google-auth>=2.0.0",
]
